+++ src/dan_lab_to_nwb/huang_2025_001617/__init__.py
@@ -1,4 +1,4 @@
-from .huang_2025_tdt_optogenetic_interface import Huang2025OptogeneticInterface
-from .huang_2025_tdt_recording_interface import Huang2025TdtRecordingInterface
+from .huang_2025_001617_optogenetic_interface import Huang2025OptogeneticInterface
+from .huang_2025_001617_recording_interface import Huang2025TdtRecordingInterface
 
-from .huang_2025_tdt_nwbconverter import Huang2025NWBConverter  # isort: skip
+from .huang_2025_001617_nwbconverter import Huang2025NWBConverter  # isort: skip
