+++ src/dan_lab_to_nwb/huang_2025_001711/__init__.py
@@ -1,4 +1,4 @@
-from .huang_2025_dlc_behavior_interface import Huang2025DlcBehaviorInterface
-from .huang_2025_dlc_ecephys_mat_interface import Huang2025DlcEcephysMatInterface
+from .huang_2025_001711_behavior_interface import Huang2025DlcBehaviorInterface
+from .huang_2025_001711_ecephys_mat_interface import Huang2025DlcEcephysMatInterface
 
-from .huang_2025_dlc_nwbconverter import Huang2025DLCNWBConverter  # isort: skip
+from .huang_2025_001711_nwbconverter import Huang2025DLCNWBConverter  # isort: skip
